board - bitmap 으로 firebase database에 upload, retrieve 완료 (storage 넣기 X) 1. 유저 이미지 (완료) 2. 게시글 리스트에서 이미지 (완료) -- bitmap type casting class 추가 3. detail activity에서 이미지 보내기
