+++ app/src/main/java/com/example/a2022swproject/mainFunction/userBoard/BoardModel/BoardRepository.java
@@ -163,24 +163,6 @@ public class BoardRepository {
 
     }
 
-//    public void getBoardImage(SingleCallBack<Result<byte[]>> callBack){
-//        final long ONE_MEGABYTE = 1024 * 1024;
-//        boardImagesRef
-//                .child("choiseoyeon0223@gmail.com_0")
-//                .getBytes(ONE_MEGABYTE)
-//                .addOnSuccessListener(new OnSuccessListener<byte[]>() {
-//                    @Override
-//                    public void onSuccess(byte[] bytes) {
-//                        callBack.onComplete(new Result.Success<byte[]>(bytes));
-//                    }
-//                })
-//                .addOnFailureListener(new OnFailureListener() {
-//                    @Override
-//                    public void onFailure(@NonNull Exception e) {
-//
-//                    }
-//                });
-//    }
 
     //게시판 불러오기
     public void getBoard(SingleCallBack<Result<ArrayList>> callBack){
